pyproject.toml
[project]
name = "mini-evermemos"
version = "0.1.0"
description = "EverMemOS-inspired memory system for AI companions"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=2.2.0",
    "pydantic>=2.0.0",
    "python-dateutil>=2.8.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pymongo>=4.6.0",
    "pymilvus>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.isort]
known-first-party = ["mini_evermemos"]

[tool.hatch.build.targets.wheel]
packages = ["src/core", "src/models", "src/prompts", "src/storage", "src/utils"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
